+++ gw_full_latest/gw_web_admin/admin/libs/python/key_clouds.py
@@ -8,6 +8,7 @@ import key_NoInternet
 import key_GpsFile
 import key_MQTT
 import key_NodeRed
+import key_TTN
 
 key_data = {'waziup_server':key_WAZIUP.waziup_server,
 		'project_name':key_WAZIUP.project_name, 
@@ -36,6 +37,7 @@ key_data = {'waziup_server':key_WAZIUP.waziup_server,
 		'nodered_organization_name':key_NodeRed.organization_name,
 		'nodered_sensor_name':key_NodeRed.sensor_name,
 		'nodered_source_list':key_NodeRed.source_list,
-		'nointernet_source_list':key_NoInternet.source_list
+		'nointernet_source_list':key_NoInternet.source_list,
+		'ttn_source_list':key_TTN.source_list
 		}
 print json.dumps(key_data)