+++ gw_full_latest/gw_web_admin/admin/libs/sh/web_shell_script.sh
@@ -417,6 +417,21 @@ then
 	fi
 fi
 
+if [ "$1" = "cloudttn_conf" ]
+then
+	###################################
+	# Configure CloudTTN
+	###################################
+
+	tmp=$(mktemp)
+	if [ $# == 3 ] 
+	then
+		jq '.clouds=([.lorawan_encrypted_clouds[]  | select(.script == "python CloudTTN.py") .'$2' = '$3'])' /home/pi/lora_gateway/clouds.json > "$tmp" && mv "$tmp" /home/pi/lora_gateway/clouds.json
+		sudo chown -R pi:pi /home/pi/lora_gateway/
+		sudo chmod +r /home/pi/lora_gateway/clouds.json
+	fi
+fi
+
 #//////////////////////////////////////////////////////////////////////////////////////////////////////
 # system GW access point
 #//////////////////////////////////////////////////////////////////////////////////////////////////////